requirements.txt
anthropic>=0.49.0
playwright>=1.40.0
notion-client>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0
rich>=13.0.0
